+++ CA1_ORM_REST_JS/src/main/java/REST/PersonResource.java
@@ -114,7 +114,7 @@ public class PersonResource {
     @Produces(MediaType.APPLICATION_JSON)
     public String getCountPersonsByZipCode(@PathParam("postalcode")String postalcode) 
     {
-         int count = pf.getCountOfPersonsByCity(postalcode);
+         Long count = pf.getCountOfPersonsByCity(postalcode);
          
          return gson.toJson(count);
     }
@@ -124,7 +124,7 @@ public class PersonResource {
     @Produces(MediaType.APPLICATION_JSON)
     public String getCountPersonsByHobby(@PathParam("hobby")String hobby) 
     {
-         int count = pf.getCountOfPersonsWithHobby(hobby);
+         Long count = pf.getCountOfPersonsWithHobby(hobby);
          
          return gson.toJson(count);
     }
